[security] Wire TLS into TCP control server
- TcpControlServer generates ephemeral TLS cert on startup via tls.rs
- listen_and_accept() wraps TCP connections with TlsAcceptor
- Handshake uses generic AsyncRead+AsyncWrite (works over TLS or plain TCP)
- read_message/send_message generified for both stream types
- Tests use new_without_tls() for plain TCP testing
- PIN type in write_status_file updated to u32 (6-digit)
- Falls back to plaintext if TLS init fails (logged as error)

+++ rust/streaming-engine/src/lib.rs
@@ -18,7 +18,7 @@ use metrics::latency::FrameTimestamps;
 
 /// Write engine status to a shared JSON file for the companion app.
 /// Path: %APPDATA%/FocusVisionPCVR/status.json (Windows)
-pub fn write_status_file(status: &str, pin: Option<u16>, latency_us: Option<u64>, fps: Option<u16>, bitrate_mbps: Option<u32>) {
+pub fn write_status_file(status: &str, pin: Option<u32>, latency_us: Option<u64>, fps: Option<u16>, bitrate_mbps: Option<u32>) {
     let dir = match dirs_next::data_dir() {
         Some(d) => d.join("FocusVisionPCVR"),
         None => return,
